requirements.txt
pymupdf
sentence-transformers
faiss-cpu
transformers
accelerate
torch
numpy
